Create package.py
In this code, there's a function called check_even() that takes a parameter num. It checks whether the number is even or odd by performing the modulo operation % with 2. If the result is 0, it means the number is even, and it prints a corresponding message. Otherwise, it prints that the number is odd.

The code assigns the value 7 to the variable number, and then calls the check_even() function with number as an argument. As a result, it will print "7 is odd."

You can modify the value of number to test different numbers and see the output.